% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ART_KF}
\alias{ART_KF}
\title{Estimating state vector by ART Kalman filter}
\usage{
ART_KF(a1, P1, T, Z, Q, H, y, lambda, S = 3L, delta_AR = 1e-08)
}
\arguments{
\item{a1}{mean of initial state vector}

\item{P1}{covariance matrix of initial state vector}

\item{T}{transition matrix of state vector}

\item{Z}{transition matrix of state vector}

\item{Q}{covariance matrix of state noise vector}

\item{H}{covariance matrix of observation noise vector}

\item{y}{observation vectors}

\item{lambda}{regularization parameter of adaptive ridge}

\item{S}{number of times to perform ridge estimation in adaptive ridge}

\item{delta_AR}{small value to prevent dividing by 0 in adaptive ridge}
}
\description{
Estimating state vector by ART Kalman filter
}
