% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TART_KF}
\alias{TART_KF}
\title{Estimating state vector by TART Kalman filter}
\usage{
TART_KF(
  a1,
  P1,
  T,
  Z,
  Q,
  H,
  y,
  lambda_init = 0,
  S = 3L,
  beta1 = 0.9,
  beta2 = 0.999,
  eta = 0.002,
  c_e = 100L,
  c_F = 5L,
  d_F = 1,
  delta_AR = 1e-08,
  delta_g = 0.01,
  delta_v = 1e-08
)
}
\arguments{
\item{a1}{mean of initial state vector}

\item{P1}{covariance matrix of initial state vector}

\item{T}{transition matrix of state vector}

\item{Z}{transition matrix of state vector}

\item{Q}{covariance matrix of state noise vector}

\item{H}{covariance matrix of observation noise vector}

\item{y}{observation vectors}

\item{lambda_init}{initial regularization parameter of adaptive ridge}

\item{S}{number of times to perform ridge estimation in adaptive ridge}

\item{beta1}{parameter of Adam}

\item{beta2}{parameter of Adam}

\item{eta}{learning rate of Adam}

\item{c_e}{interval over which the average is taken for M_t in TART-KF}

\item{c_F}{interval over which the average is taken for F_t in TART-KF}

\item{d_F}{rate of lambda control by Kalman filter}

\item{delta_AR}{small value to prevent dividing by 0 in adaptive ridge}

\item{delta_g}{value to approximate derivative of L_t for Adam}

\item{delta_v}{small value to prevent dividing by 0 in Adam}
}
\description{
Estimating state vector by TART Kalman filter
}
