% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TART_KF}
\alias{TART_KF}
\title{Estimating state vector by TART Kalman filter}
\usage{
TART_KF(
  a1,
  P1,
  T,
  Z,
  Sigma_eta,
  Sigma_epsilon,
  y,
  lambda_init = 0,
  S = 3L,
  beta1 = 0.9,
  beta2 = 0.999,
  gamma = 0.002,
  c_e = 100L,
  c_F = 5L,
  d_F = 1,
  delta_AR = 1e-08,
  delta_g = 0.01,
  delta_v = 1e-08
)
}
\arguments{
\item{a1}{p * 1 mean of initial state vector}

\item{P1}{p * p covariance matrix of initial state vector}

\item{T}{p * p * n (or p * p * n_T, if T has an n_T cycle) array giving transition matrix in state equation for each time}

\item{Z}{d * p * n (or p * p * n_Z, if Z has an n_Z cycle) array giving transition matrix in observation equation for each time}

\item{Sigma_eta}{p * p * n (or p * p * n_Sigma_eta, if Sigma_eta has an n_Sigma_eta cycle) array giving covariance matrix of state noise vector for each time}

\item{Sigma_epsilon}{d * d * n (or d * d * n_Sigma_epsilon, if Sigma_epsilon has an n_Sigma_epsilon cycle) array giving covariance matrix of observation noise vector for each time}

\item{y}{d * n observation vector for each time}

\item{lambda_init}{initial regularization parameter of adaptive ridge}

\item{S}{number of times to perform ridge estimation in adaptive ridge}

\item{beta1}{parameter of Adam}

\item{beta2}{parameter of Adam}

\item{gamma}{learning rate of Adam}

\item{c_e}{interval over which the average is taken for M_t in TART-KF}

\item{c_F}{interval over which the average is taken for F_t in TART-KF}

\item{d_F}{rate of lambda control by Kalman filter}

\item{delta_AR}{small value to prevent dividing by 0 in adaptive ridge}

\item{delta_g}{value to approximate derivative of L_t for Adam}

\item{delta_v}{small value to prevent dividing by 0 in Adam}
}
\description{
Estimating state vector by TART Kalman filter
}
